+++ metadata-ingestion/src/datahub/ingestion/autogenerated/connector_registry/datahub.json
@@ -1,6 +1,6 @@
 {
-  "generated_by": "metadata-ingestion/scripts/connector_registry.py",
-  "package": "datahub",
+  "generated_at": "2025-12-01T00:13:13.933940+00:00",
+  "generated_by": "metadata-ingestion/scripts/capability_summary.py",
   "plugin_details": {
     "abs": {
       "capabilities": [
@@ -43,11 +43,23 @@
       "platform_name": "ABS Data Lake",
       "support_status": "INCUBATING"
     },
-    "aerospike": {
+    "airbyte": {
       "capabilities": [
         {
+          "capability": "LINEAGE_FINE",
+          "description": "Enabled by default",
+          "subtype_modifier": null,
+          "supported": true
+        },
+        {
           "capability": "DELETION_DETECTION",
-          "description": "Enabled by default via stateful ingestion",
+          "description": "Enabled by default when stateful ingestion is turned on.",
+          "subtype_modifier": null,
+          "supported": true
+        },
+        {
+          "capability": "TAGS",
+          "description": "Requires recipe configuration",
           "subtype_modifier": null,
           "supported": true
         },
@@ -58,16 +70,16 @@
           "supported": true
         },
         {
-          "capability": "SCHEMA_METADATA",
+          "capability": "LINEAGE_COARSE",
           "description": "Enabled by default",
           "subtype_modifier": null,
           "supported": true
         }
       ],
-      "classname": "datahub.ingestion.source.aerospike.AerospikeSource",
-      "platform_id": "aerospike",
-      "platform_name": "Aerospike",
-      "support_status": "TESTING"
+      "classname": "datahub.ingestion.source.airbyte.source.AirbyteSource",
+      "platform_id": "airbyte",
+      "platform_name": "Airbyte",
+      "support_status": "CERTIFIED"
     },
     "athena": {
       "capabilities": [
